requirements.txt
Django==1.9.7
django-filter==0.13.0
django-graphiql==0.4.4
djangorestframework==3.3.3
graphene==0.10.2
graphql-core==0.5.2
graphql-django-view==1.3
graphql-relay==0.4.3
iso8601==0.1.11
promise==0.4.1
singledispatch==3.4.0.3
six==1.10.0
